requirements.txt
opencv-python
numpy
matplotlib
Pillow
keras-facenet
tensorflow
keras
scipy
